+++ packages/inversify-bundler-express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rym-lib/inversify-bundler-express
 
+## 1.2.4
+
+### Patch Changes
+
+- fix query-module filter type missmatch
+- Updated dependencies
+  - @rym-lib/inversify-bundler@1.2.4
+
 ## 1.2.3
 
 ### Patch Changes