+++ packages/nakadachi-interactor-mixin-validator/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rym-lib/nakadachi-interactor-mixin-validator
 
+## 1.2.4
+
+### Patch Changes
+
+- fix query-module filter type missmatch
+- Updated dependencies
+  - @rym-lib/exception@1.2.4
+  - @rym-lib/inversify-bundler@1.2.4
+  - @rym-lib/nakadachi@1.2.4
+
 ## 1.2.3
 
 ### Patch Changes