+++ packages/rdb-command/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rym-lib/rdb-command
 
+## 1.2.4
+
+### Patch Changes
+
+- fix query-module filter type missmatch
+
 ## 1.2.3
 
 ### Patch Changes